+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 
 # 0.2.10 (TBD)
+### Improvements
+- [quantum] - Bump Ingenii Azure Quantum package to 0.1.0.
 
 # 0.2.9 (2022-03-22)